Switch to Nginx

+++ start.sh
@@ -1,9 +1,7 @@
 #!/bin/bash
-if [ ! -f .env ]; then
-  cp .env.save .env
-fi
+if [ ! -f .env ]; then cp .env.save .env; fi
 php artisan key:generate --force || true
 php artisan migrate --force || true
-sed -i "s/Listen 80/Listen ${PORT}/g" /etc/apache2/ports.conf
-sed -i "s/*:80/*:${PORT}/g" /etc/apache2/sites-available/000-default.conf
-exec apache2-foreground
+echo "server { listen ${PORT}; root /var/www/html/public; index index.php; location / { try_files \$uri \$uri/ /index.php?\$query_string; } location ~ \.php$ { fastcgi_pass unix:/run/php/php8.2-fpm.sock; fastcgi_index index.php; include fastcgi_params; fastcgi_param SCRIPT_FILENAME \$document_root\$fastcgi_script_name; } }" > /etc/nginx/sites-available/default
+php-fpm8.2
+exec nginx -g 'daemon off;'
